# Convert git deps to local, faster tests