--- a/tests/functional/schema_tests/test_schema_v2_tests.py
+++ b/tests/functional/schema_tests/test_schema_v2_tests.py
@@ -3,6 +3,8 @@ import os
 import re
 
 from dbt.tests.util import run_dbt, write_file
+from dbt.tests.fixtures.project import write_project_files
+from tests.fixtures.dbt_integration_project import dbt_integration_project  # noqa: F401
 from tests.functional.schema_tests.fixtures import (  # noqa: F401
     wrong_specification_block,
     test_context_where_subq_models,
@@ -386,7 +388,8 @@ class TestHooksForWhich:
 
 class TestCustomSchemaTests:
     @pytest.fixture(scope="class", autouse=True)
-    def setUp(self, project):
+    def setUp(self, project, project_root, dbt_integration_project):  # noqa: F811
+        write_project_files(project_root, "dbt_integration_project", dbt_integration_project)
         project.run_sql_file(os.path.join(project.test_data_dir, "seed.sql"))
 
     @pytest.fixture(scope="class")
@@ -397,8 +400,7 @@ class TestCustomSchemaTests:
                     "local": "./local_dependency",
                 },
                 {
-                    "git": "https://github.com/dbt-labs/dbt-integration-project",
-                    "revision": "dbt/1.0.0",
+                    "local": "./dbt_integration_project",
                 },
             ]
         }