resolves #5514

Copy https://github.com/dbt-labs/dbt-integration-project into a set of local fixtures that we can more quickly set up in functional tests.

### Before: 64.51s

```
$ python3 -m pytest \
> tests/functional/context_methods/test_cli_vars.py::TestCLIVarsPackages \
> tests/functional/graph_selection/test_schema_test_graph_selection.py \
> tests/functional/schema_tests/test_schema_v2_tests.py::TestCustomSchemaTests
============================================= test session starts ==============================================
platform darwin -- Python 3.9.12, pytest-7.0.1, pluggy-1.0.0
rootdir: /Users/jerco/dev/product/dbt-core, configfile: pytest.ini
plugins: xdist-2.5.0, forked-1.4.0, csv-3.0.0, logbook-1.2.0, flaky-3.7.0, mock-3.8.2, dotenv-0.5.2, cov-3.0.0
collected 15 items

tests/functional/context_methods/test_cli_vars.py .                                                      [  6%]
tests/functional/graph_selection/test_schema_test_graph_selection.py .............                       [ 93%]
tests/functional/schema_tests/test_schema_v2_tests.py .                                                  [100%]

======================================== 15 passed in 64.51s (0:01:04) =========================================
```

### After: 15.97s
```
$ python3 -m pytest \
> tests/functional/context_methods/test_cli_vars.py::TestCLIVarsPackages \
> tests/functional/graph_selection/test_schema_test_graph_selection.py \
> tests/functional/schema_tests/test_schema_v2_tests.py::TestCustomSchemaTests
============================================= test session starts ==============================================
platform darwin -- Python 3.9.12, pytest-7.0.1, pluggy-1.0.0
rootdir: /Users/jerco/dev/product/dbt-core, configfile: pytest.ini
plugins: xdist-2.5.0, forked-1.4.0, csv-3.0.0, logbook-1.2.0, flaky-3.7.0, mock-3.8.2, dotenv-0.5.2, cov-3.0.0
collected 15 items

tests/functional/context_methods/test_cli_vars.py .                                                      [  6%]
tests/functional/graph_selection/test_schema_test_graph_selection.py .............                       [ 93%]
tests/functional/schema_tests/test_schema_v2_tests.py .                                                  [100%]

============================================= 15 passed in 15.97s ==============================================
```

### "Mach Speed": 12.52s

@dbeatty10 iykyk  :)

```
$ python3 -m pytest \
> tests/functional/context_methods/test_cli_vars.py::TestCLIVarsPackages \
> tests/functional/graph_selection/test_schema_test_graph_selection.py \
> tests/functional/schema_tests/test_schema_v2_tests.py::TestCustomSchemaTests
======================================== test session starts =========================================
platform darwin -- Python 3.9.12, pytest-7.0.1, pluggy-1.0.0
rootdir: /Users/jerco/dev/product/dbt-core, configfile: pytest.ini
plugins: xdist-2.5.0, forked-1.4.0, csv-3.0.0, logbook-1.2.0, flaky-3.7.0, mock-3.8.2, dotenv-0.5.2, cov-3.0.0
collected 15 items

tests/functional/context_methods/test_cli_vars.py .                                            [  6%]
tests/functional/graph_selection/test_schema_test_graph_selection.py .............             [ 93%]
tests/functional/schema_tests/test_schema_v2_tests.py .                                        [100%]

======================================== 15 passed in 12.52s =========================================
```

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-core/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] I have run this code in development and it appears to resolve the stated issue
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have [opened an issue to add/update docs](https://github.com/dbt-labs/docs.getdbt.com/issues/new/choose), or docs changes are not required/relevant for this PR
- ~I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-core/blob/main/CONTRIBUTING.md#Adding-CHANGELOG-Entry)~